+++ src/hotkeys.js
@@ -2,67 +2,58 @@ import CONSTANTS from "./constants/constants.js";
 import * as Utilities from "./helpers/utilities.js";
 import PrivateAPI from "./API/private-api.js";
 
+const HOTKEYS = {
+  FORCE_DEFAULT_SHEET: "force-open-item-pile-inventory",
+  DROP: "force-drop-item",
+  DROP_ONE: "force-drop-one-item"
+}
+
 export const hotkeyActionState = {
   get openPileInventory() {
-    return (!hotkeyState.ctrlDown && !game.settings.get(CONSTANTS.MODULE_NAME, "invertSheetOpen"))
-      || (hotkeyState.ctrlDown && game.settings.get(CONSTANTS.MODULE_NAME, "invertSheetOpen"));
+    const down = window.keyboard.downKeys.has(game.keybindings.get(CONSTANTS.MODULE_NAME, HOTKEYS.FORCE_DEFAULT_SHEET)[0].key);
+    return (
+      (!down && !game.settings.get(CONSTANTS.MODULE_NAME, "invertSheetOpen"))
+      ||
+      (down && game.settings.get(CONSTANTS.MODULE_NAME, "invertSheetOpen"))
+    );
+  },
+
+  get forceDropItem() {
+    return window.keyboard.downKeys.has(game.keybindings.get(CONSTANTS.MODULE_NAME, HOTKEYS.DROP)[0].key);
+  },
+
+  get forceDropOneItem() {
+    return window.keyboard.downKeys.has(game.keybindings.get(CONSTANTS.MODULE_NAME, HOTKEYS.DROP_ONE)[0].key);
   }
 }
 
-export const hotkeyState = {
-  ctrlDown: false,
-  altDown: false,
-  shiftDown: false
-}
-
 export function registerHotkeysPre() {
-  
-  game.keybindings.register(CONSTANTS.MODULE_NAME, "force-open-item-pile-inventory", {
-    name: "Force open inventory",
-    uneditable: [
+
+  game.keybindings.register(CONSTANTS.MODULE_NAME, HOTKEYS.FORCE_DEFAULT_SHEET, {
+    name: "Force open inventory modifier",
+    editable: [
       { key: "ControlLeft" },
-    ],
-    onDown: () => {
-      hotkeyState.ctrlDown = true;
-    },
-    onUp: () => {
-      hotkeyState.ctrlDown = false;
-    },
-    reservedModifiers: ["SHIFT", "ALT"]
+    ]
   });
-  
-  game.keybindings.register(CONSTANTS.MODULE_NAME, "force-drop-item", {
-    name: "Force drop item (GM only)",
-    uneditable: [
+
+  game.keybindings.register(CONSTANTS.MODULE_NAME, HOTKEYS.DROP, {
+    name: "Force drop item (GM only) modifier",
+    editable: [
       { key: "ShiftLeft" },
-    ],
-    onDown: () => {
-      hotkeyState.shiftDown = true;
-    },
-    onUp: () => {
-      hotkeyState.shiftDown = false;
-    },
-    reservedModifiers: ["ALT", "CONTROL"]
+    ]
   });
-  
-  game.keybindings.register(CONSTANTS.MODULE_NAME, "force-drop-one-item", {
-    name: "Force drop one item",
-    uneditable: [
+
+  game.keybindings.register(CONSTANTS.MODULE_NAME, HOTKEYS.DROP_ONE, {
+    name: "Force drop one item modifier",
+    editable: [
       { key: "AltLeft" },
-    ],
-    onDown: () => {
-      hotkeyState.altDown = true;
-    },
-    onUp: () => {
-      hotkeyState.altDown = false;
-    },
-    reservedModifiers: ["SHIFT", "CONTROL"]
+    ]
   });
-  
+
 }
 
 export function registerHotkeysPost() {
-  
+
   if (!game.user.isGM) {
     let clicked = false;
     window.addEventListener("mousedown", (event) => {
@@ -72,7 +63,7 @@ export function registerHotkeysPost() {
       const hover = document.elementFromPoint(event.clientX, event.clientY);
       if (!hover || (hover.id !== "board")) return;
       if (event.button !== 0) return;
-      
+
       const pos = canvas.app.renderer.plugins.interaction.mouse.getLocalPosition(canvas.app.stage);
       const tokens = Utilities.getTokensAtLocation(pos)
         .filter(token => {
@@ -83,12 +74,12 @@ export function registerHotkeysPost() {
       if (!tokens.length) return;
       tokens.sort((a, b) => b.zIndex - a.zIndex);
       const token = Utilities.getDocument(tokens[0]);
-      
+
       if (clicked === token) {
         clicked = false;
         return PrivateAPI._itemPileClicked(token);
       }
-      
+
       clicked = token;
       setTimeout(() => {
         clicked = false;