+++ src/libwrapper.js
@@ -1,5 +1,4 @@
 import CONSTANTS from "./constants/constants.js";
-import { hotkeyActionState } from "./hotkeys.js";
 import PrivateAPI from "./API/private-api.js";
 import * as PileUtilities from "./helpers/pile-utilities.js";
 import * as Helpers from "./helpers/helpers.js";
@@ -7,7 +6,7 @@ import * as Helpers from "./helpers/helpers.js";
 export default function registerLibwrappers() {
 
   libWrapper.register(CONSTANTS.MODULE_NAME, 'Token.prototype._onClickLeft2', function (wrapped, ...args) {
-    if (PileUtilities.isValidItemPile(this.document) && hotkeyActionState.openPileInventory) {
+    if (PileUtilities.isValidItemPile(this.document) && !keyboard.downKeys.has("ControlLeft")) {
       return PrivateAPI._itemPileClicked(this.document);
     }
     return wrapped(...args);
@@ -32,7 +31,7 @@ export default function registerLibwrappers() {
   }, "MIXED");
 
   libWrapper.register(CONSTANTS.MODULE_NAME, `ActorSheet.prototype.render`, function (wrapped, forced, options, ...args) {
-    const renderItemPileInterface = forced && !options?.bypassItemPiles && PileUtilities.isValidItemPile(this.document) && hotkeyActionState.openPileInventory;
+    const renderItemPileInterface = forced && !options?.bypassItemPiles && PileUtilities.isValidItemPile(this.document) && !keyboard.downKeys.has("ControlLeft");
     if (this._state > Application.RENDER_STATES.NONE) {
       if (renderItemPileInterface) {
         wrapped(forced, options, ...args)